requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Deep learning
tensorflow>=2.8.0
keras>=2.8.0

# Graph Neural Networks
spektral>=1.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Utilities
tqdm>=4.62.0

